+++ bin/src/model/model.dart
@@ -7,16 +7,13 @@ void insertModel({required ArgResults from}) {
       throw '\x1B[0mAvailable commands :';
     } else {
       String dir = from['input']!;
-      List<File> files =
-          Directory(dir.endsWith('/') ? dir.substring(0, dir.length - 1) : dir)
-              .listSync(recursive: true)
-              .whereType<File>()
-              .where((e) =>
-                  e.path.split('.').last.trim() == 'dart' &&
-                  e
-                      .readAsStringSync()
-                      .contains(RegExp(r'(?=@Model)|(?=@model)')))
-              .toList();
+      List<File> files = Directory(dir.endsWith('/') ? dir.substring(0, dir.length - 1) : dir)
+          .listSync(recursive: true)
+          .whereType<File>()
+          .where((e) =>
+              e.path.split('.').last.trim() == 'dart' &&
+              e.readAsStringSync().contains(RegExp(r'(?=@Model)|(?=@model)')))
+          .toList();
       for (int index = 0; index < files.length; index++) {
         ModelParser.run(files.elementAt(index));
         stdout.write('\rUpdating model ${index + 1}/${files.length} ');
@@ -120,9 +117,9 @@ extension StringListExtension on List<String> {
 
 class ModelParser {
   const ModelParser({
-    this.doc,
     required this.end,
     required this.name,
+    required this.test,
     required this.begin,
     required this.toJSON,
     required this.copyWith,
@@ -133,41 +130,74 @@ class ModelParser {
   final bool toJSON;
   final bool copyWith;
   final bool fromJSON;
-  final String? doc;
   final bool immutable;
+  final bool test;
   final int begin;
   final int end;
 
   static File run(File file) {
-    var data = file
-        .readAsStringSync()
-        .split(RegExp(r'(?=@Model)|(?=@model)'))
-        .map((e) {
+    var data = file.readAsStringSync().split(RegExp(r'(?=@Model)|(?=@model)')).map((e) {
       if (e.trim().startsWith(RegExp(r'(?=@Model)|(?=@model)'))) {
         final value = e.split('\n');
-        bool noToJSON =
-            value.contain('toJSON: false', sources: ['@Model(...)']);
-        bool noCopyWith =
-            value.contain('copyWith: false', sources: ['@Model(...)']);
-        bool noFromJSON =
-            value.contain('fromJSON: false', sources: ['@Model(...)']);
-        bool noImmutable =
-            value.contain('immutable: false', sources: ['@Model(...)']);
+        bool noToJSON = value.contain('toJSON: false', sources: ['@Model(...)']);
+        bool noCopyWith = value.contain('copyWith: false', sources: ['@Model(...)']);
+        bool noFromJSON = value.contain('fromJSON: false', sources: ['@Model(...)']);
+        bool noImmutable = value.contain('immutable: false', sources: ['@Model(...)']);
+        bool noTest = value.contain('test: false', sources: ['@Model(...)']);
         final range = value.range(['@Model(...)', '@model']);
         List<VariableParser> variables = VariableParser.fromString(e);
         StringBuffer buffer = StringBuffer();
         final model = ModelParser(
           end: value.length,
           begin: 0,
-          name: RegExp(r'(?<=class\s).+?\w+')
-              .allMatches(value[range.last + 1])
-              .first
-              .group(0)!,
+          test: !noTest,
+          name: RegExp(r'(?<=class\s).+?\w+').allMatches(value[range.last + 1]).first.group(0)!,
           toJSON: !noToJSON,
           copyWith: !noCopyWith,
           fromJSON: !noFromJSON,
           immutable: !noImmutable,
         );
+
+        if (model.test) {
+          value.replace('static ${model.name} get test...}', replacement: (exist) {
+            buffer.clear();
+            for (var variable in variables) {
+              buffer.write(model.immutable ? '\n\t\t\t${variable.name}: ' : '\n\t\t\t..${variable.name} = ');
+              if (variable.defaultsTo != null) {
+                buffer.write('${variable.defaultsTo}');
+              } else {
+                switch (variable.type) {
+                  case 'String':
+                    buffer.write("'This is a sample ${variable.name}'");
+                    break;
+                  case 'int':
+                    buffer.write('0');
+                    break;
+                  case 'double':
+                    buffer.write('1.0');
+                    break;
+                  case 'bool':
+                    buffer.write('false');
+                    break;
+                  default:
+                    if (variable.type.startsWith('List')) {
+                      buffer.write('const []');
+                      break;
+                    } else {
+                      buffer.write('${variable.type}.test');
+                      break;
+                    }
+                }
+              }
+
+              buffer.write(model.immutable ? ',' : '');
+            }
+            return '${exist ? '' : '\n'}'
+                '\tstatic ${model.name} get test {\n'
+                '${model.immutable ? '\t\treturn ${model.name}($buffer\n\t\t)' : '\t\treturn ${model.name}()$buffer\n\t\t'}'
+                ';\n\t}';
+          });
+        }
 
         if (model.copyWith) {
           value.replace('\t${model.name} copyWith...}', replacement: (exist) {
@@ -179,8 +209,7 @@ class ModelParser {
             if (model.immutable) {
               buffer.write('\t\treturn ${model.name}(\n');
               for (var variable in variables) {
-                buffer.write(
-                    '\t\t\t${variable.name}: ${variable.name} ?? this.${variable.name},\n');
+                buffer.write('\t\t\t${variable.name}: ${variable.name} ?? this.${variable.name},\n');
               }
               buffer.write('\t\t);\n');
             } else {
@@ -203,7 +232,7 @@ class ModelParser {
             buffer.clear();
             for (var variable in variables) {
               buffer.write(
-                  "\t\t'${variable.key ?? variable.name}': ${variable.name}${variable.toJSON ? '.toJSON' : ''}, \n");
+                  "\t\t'${variable.key ?? variable.name}': ${variable.name}${variable.toJSON ? '${variable.nullable ? '?' : ''}.toJSON' : ''}, \n");
             }
             return '${exist ? '' : '\n\t@override\n'}'
                 '\tJSON get toJSON => {\n'
@@ -214,15 +243,18 @@ class ModelParser {
         }
 
         if (model.fromJSON) {
-          value.replace('static ${model.name} fromJSON...}',
-              replacement: (exist) {
+          value.replace('static ${model.name} fromJSON...}', replacement: (exist) {
             buffer.clear();
             for (var variable in variables) {
+              String prefix = variable.nullable ? 'value.maybeOf' : 'value.of';
+              String suffixNullable =
+                  variable.nullable && variable.defaultsTo != null ? ' ?? ${variable.defaultsTo}' : '';
+              String suffix = !variable.nullable && variable.defaultsTo != null ? ', ${variable.defaultsTo}' : '';
               String content = variable.fromJSON
                   ? variable.type.contains('List')
-                      ? "value.of<List<JSON>>('${variable.key ?? variable.name}').map(${variable.type.replaceAll(RegExp(r'.+?(?<=<)|(?=>).+?'), '')}.fromJSON).toList()"
-                      : "${variable.type}.fromJSON(value.of<JSON>('${variable.key ?? variable.name}'))"
-                  : "value.of<${variable.type}>('${variable.key ?? variable.name}')";
+                      ? "$prefix<List<JSON>>('${variable.key ?? variable.name}'$suffix).map(${variable.type.replaceAll(RegExp(r'.+?(?<=<)|(?=>).+?'), '')}.fromJSON).toList()$suffixNullable"
+                      : "${variable.type}.fromJSON($prefix<JSON>('${variable.key ?? variable.name}'$suffix)$suffixNullable${suffixNullable.isNotEmpty ? '.toJSON' : ''})"
+                  : "$prefix<${variable.type}>('${variable.key ?? variable.name}'$suffix)$suffixNullable";
               if (model.immutable) {
                 buffer.write("\n\t\t\t${variable.name}: $content,");
               } else {
@@ -244,8 +276,7 @@ class ModelParser {
       }
       return e;
     }).join('');
-    if (data.contains('immutable: false') &&
-        !data.contains('must_be_immutable')) {
+    if (data.contains('immutable: false') && !data.contains('must_be_immutable')) {
       data = '// ignore_for_file: must_be_immutable\n\n$data';
     }
     file.writeAsStringSync(data);
@@ -256,11 +287,11 @@ class ModelParser {
   String toString() => '$runtimeType('
       'begin: $begin, '
       'end: $end, '
-      'doc: $doc, '
       'name: $name, '
       'to_json: $toJSON, '
       'from_json: $fromJSON, '
       'copy_with: $copyWith, '
+      'test: $test, '
       'immutable: $immutable)';
 }
 
@@ -271,12 +302,16 @@ class VariableParser {
     required this.name,
     this.toJSON = false,
     this.fromJSON = false,
+    this.nullable = false,
+    this.defaultsTo,
   });
   final String? key;
   final String name;
   final String type;
+  final bool nullable;
   final bool toJSON;
   final bool fromJSON;
+  final String? defaultsTo;
 
   static List<VariableParser> fromString(String data) {
     List<VariableParser> objects = [];
@@ -284,24 +319,24 @@ class VariableParser {
     for (int index = 0; index < value.length; index++) {
       if (value[index].startsWith(RegExp(r'(?=@Variable)|(?=@variable)'))) {
         final values = value[index].split('\n');
-        final range = values.range(['@Variable...)', '@variable']);
+        final range = values.range(['@Variable...)\n', '@variable']);
         String line = values[range.last + 1];
         final key = RegExp(r'''(?<=name:).*(?<='|")''')
             .stringMatch(values[range.last])
             ?.trim()
             .replaceAll(RegExp(r'''('|")'''), '');
-        bool toJSON =
-            values.contain('toJSON: true', sources: ['@Variable(...)']);
-        bool fromJSON =
-            values.contain('fromJSON: true', sources: ['@Variable(...)']);
-        List<String> variable = line
-            .replaceAll(RegExp(r'final |const |static |;'), '')
-            .trim()
-            .split(' ');
+        bool toJSON = values.contain('toJSON: true', sources: ['@Variable(...)']);
+        bool fromJSON = values.contain('fromJSON: true', sources: ['@Variable(...)']);
+        List<String> variable = line.replaceAll(RegExp(r'final |const |static |;'), '').trim().split(' ');
+        var defaultsTo = RegExp(r'(?<=defaultsTo:).*(?=, toJSON|, fromJSON|, name|,.*?\)\\n|\))')
+            .stringMatch(values[range.last])
+            ?.trim();
         final model = VariableParser(
             key: key,
-            type: variable[0],
+            defaultsTo: defaultsTo,
+            type: variable[0].trim().replaceAll('?', ''),
             name: variable[1],
+            nullable: variable[0].trim().endsWith('?'),
             toJSON: toJSON,
             fromJSON: fromJSON);
         objects.add(model);
@@ -316,6 +351,8 @@ class VariableParser {
         'key: $key, '
         'name; $name, '
         'type: $type, '
+        'nullable: $nullable, '
+        'defaults_to; $defaultsTo, '
         'toJSON: $toJSON, '
         'fromJSON: $fromJSON)';
   }