+++ lib/src/d_models.dart
@@ -32,13 +32,20 @@ class DModel extends Equatable {
   /// ```
   static DModel fromJSON(JSON value) => const DModel();
 
-  @override
-  List<Object?> get props =>
-      toJSON.entries.map((e) => '${e.key}: ${e.value}').toList();
+  /// A Dummy test model value.
+  ///
+  /// ```dart
+  /// DModel test = Dmodel.test;
+  /// ```
+  static DModel get test {
+    return const DModel();
+  }
 
   @override
-  String toString() =>
-      toJSON.entries.map((e) => '${e.key}: ${e.value}').toString();
+  List<Object?> get props => toJSON.entries.map((e) => '${e.key}: ${e.value}').toList();
+
+  @override
+  String toString() => toJSON.entries.map((e) => '${e.key}: ${e.value}').toString();
 }
 
 /// Basic model in root of every `Response`, containing [success] status,
@@ -59,8 +66,7 @@ class ResponseModel<T extends DModel> extends DModel {
     this.data,
   });
 
-  static ResponseModel fromJSON<T extends DModel>(JSON value,
-      {T Function(JSON value)? data}) {
+  static ResponseModel fromJSON<T extends DModel>(JSON value, {T Function(JSON value)? data}) {
     return ResponseModel(
       success: value.of('success'),
       message: value.of('message'),