+++ CHANGELOG.md
@@ -69,3 +69,6 @@
 * Remove GoRouter Extension
 * Add DLogWidget
 * Display scanned image file in doc.
+* Add nullable scanner in Model generator
+* Add Test model generation in Model generator and DModel
+* Add defaultsTo in Variable annotation